Implemented mappool format generator

+++ src/routes/(main)/tournaments/[id]/staff-dashboard/mappools/+page.server.ts
@@ -4,6 +4,7 @@ import prisma from "$lib/prisma";
 import { StatusCodes } from "$lib/StatusCodes";
 import { parseFormData } from "parse-nested-form-data";
 import vine, { errors } from "@vinejs/vine";
+import { ModInGameNames, ModList, Mods } from "$lib/ModEnums";
 
 export const load: PageServerLoad = async () => {
     return {};
@@ -51,7 +52,7 @@ export const actions: Actions = {
             return {
                 status: StatusCodes.CREATED,
                 message: 'New round created.',
-                data: newRound
+                roundId: newRound.id
             }
         }
 
@@ -60,15 +61,49 @@ export const actions: Actions = {
         });    
     },
 
-    delete_mappool: async ({ request }) => {
+    delete_mappool: async ({ request, params }) => {
         const formData = parseFormData(await request.formData());
         const id = parseInt(String(formData.id));
 
+        // Retrieve round before deletion
+        const round = await prisma.round.findFirst({
+            where: {
+                id
+            },
+            select: {
+                id: true,
+                mappoolId: true,
+                tournamentId: true
+            }
+        });
+
+        if (!round) {
+            return fail(StatusCodes.BAD_REQUEST, {
+                message: 'This round does not exist.'
+            })
+        }
+
+        // Validate the tournament ID is correct
+        // (prevents the changing of round ID on the client-side being able to delete another tournament's round)
+        if (parseInt(String(params.id)) != round.tournamentId) {
+            return fail(StatusCodes.BAD_REQUEST, {
+                message: 'Invalid tournament ID.'
+            });
+        }
+
+        // Delete the round's mappool
+        await prisma.mappool.delete({
+            where: {
+                id: round.mappoolId ?? undefined
+            }
+        });
+
+        // Finally, detele the round
         await prisma.round.delete({
             where: {
                 id
             }
-        });
+        })
     },
 
     update_mappool: async ({ request }) => {
@@ -103,28 +138,137 @@ export const actions: Actions = {
 		}
     },
 
-    // This approach is significantly more user friendly but impossible due to
-    //  MapInPool.Map and .mapId being mandatory attributes
+
     generate_mappool: async ({ request }) => {
         // Takes the mod pools and number of slots per modpool
         const formData = parseFormData(await request.formData());
-        // console.log(JSON.stringify(formData));
+        let roundId = formData.round_id;
+        
+        if (roundId == null || roundId == undefined) {
+            return fail(StatusCodes.BAD_REQUEST, {
+                message: 'No mappool id provided.'
+            });
+            // Could alternatively create a new mappool but this could be abused
+        }
+
+        roundId = Number(roundId);
+        if (isNaN(roundId)) {
+            return fail(StatusCodes.BAD_REQUEST, {
+                message: 'Invalid round ID provided.'
+            });
+        }
+
+        // Find mappool through round
+        const round = await prisma.round.findUnique({
+            where: {
+                id: roundId
+            },
+            select: {
+                mappoolId: true
+            }
+        });
+
+        let mappoolId = round?.mappoolId;
+
+        if (!mappoolId) {
+            // Create a blank mappool for this round
+            const newMappool = await prisma.mappool.create({
+                data: {
+                    global: false,
+                    Round: {
+                        connect: {
+                            id: roundId
+                        }
+                    },
+                    // tournament and round related attributes will stay null for now (like test data)
+                    // (this is scuffed)
+                    Maps: {
+                        create: []
+                    }
+                }
+            });
+
+            mappoolId = newMappool.id;
+        }
 
         // Generate empty MapInPool entries for each slot
+        const mapIds: string[] = [];
         for (const key in formData) {
-            const nOfSlots = Number(formData[key]);
-            
-            if (!isNaN(nOfSlots)) {
-                for (let slot = 1; slot <= nOfSlots; slot++) {
-                    await prisma.mapInPool.create({
-                        data: {
-                            identifier: `${key}${slot}`,
-                            mods: key,
-
-                        }
-                    });
-                }
+            if (key == "round_id") {
+                continue;
             }
+
+            let nOfSlots = Number(formData[key]);
+            if (isNaN(nOfSlots)) {
+                nOfSlots = 0;
+            }
+            
+            // Validate the key is a valid mod
+            const modPriority = ModList.indexOf(key as Mods);
+            if (modPriority == -1) {
+                return fail(StatusCodes.BAD_REQUEST, {
+                    message: 'Invalid mod provided.'
+                });
+            }
+
+            for (let slot = 1; slot <= nOfSlots; slot++) {
+                const identifier = `${key}${slot}`;
+                mapIds.push(identifier);
+
+                await prisma.mapInPool.upsert({
+                    create: {
+                        identifier,
+                        mods: ModInGameNames[key],
+                        modPriority: ModList.indexOf(key as Mods),
+                        Mappool: {
+                            connect: {
+                                id: mappoolId
+                            }
+                        },
+                        InMatches: {
+                            create: []
+                        }
+                    },
+                    update: {},
+                    where: {
+                        identifier_mappoolId: {
+                            identifier,
+                            mappoolId
+                        },
+                    }
+                });
+            }
+        }
+
+        // Delete all maps from the mappool that were not generated/updated by this action
+        const allMapsInPool = await prisma.mapInPool.findMany({
+            where: {
+                mappoolId
+            },
+            select: {
+                identifier: true,
+                mappoolId: true
+            }
+        });
+
+        allMapsInPool.forEach(async mapInPool =>{
+            if (!mapIds.some(m => m == mapInPool.identifier)) {
+                await prisma.mapInPool.delete({
+                    where: {
+                        identifier_mappoolId: {
+                            identifier: mapInPool.identifier,
+                            mappoolId: mapInPool.mappoolId
+                        }
+                    }
+                });
+            }
+        });
+
+
+        return {
+            status: StatusCodes.CREATED,
+            message: 'Mappool generated.',
+            roundId: roundId
         }
     }
 }
